+++ src/Component/MainPage/index.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function MainPage() {
+  return (
+    <div>Hello World</div>
+  )
+}
